Fix the bot action for more strict data struct

+++ tools/bot-actions/BotTargetListNextPos.cpp
@@ -31,16 +31,16 @@ std::pair<uint8_t, uint8_t> BotTargetList::getNextPosition(const MapServerMini::
             case ActionsBotInterface::GlobalTarget::Fight:
                 for(auto it = blockObject->botsFight.begin();it!=blockObject->botsFight.cend();++it)
                 {
-                    const std::vector<uint32_t> &botsFightList=it->second;
-                    if(vectorcontainsAtLeastOne(botsFightList,target.extra))
+                    const std::vector<uint16_t> &botsFightList=it->second;
+                    if(vectorcontainsAtLeastOne(botsFightList,(uint16_t)target.extra))
                         return it->first;
                 }
             break;
             case ActionsBotInterface::GlobalTarget::Shop:
                 for(auto it = blockObject->shops.begin();it!=blockObject->shops.cend();++it)
                 {
-                    const std::vector<uint32_t> &shops=it->second;
-                    if(vectorcontainsAtLeastOne(shops,target.extra))
+                    const std::vector<uint16_t> &shops=it->second;
+                    if(vectorcontainsAtLeastOne(shops,(uint16_t)target.extra))
                         return it->first;
                 }
             break;
