+++ tools/bot/actions/MapServerMini.h
@@ -104,9 +104,9 @@ public:
 
         //things into each zone
         std::map<std::pair<uint8_t,uint8_t>,ItemOnMap/*,pairhash*/> pointOnMap_Item;
-        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint32_t>,pairhash> botsFight;
-        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint32_t>,pairhash> botsFightTrigger;//trigger line in front of bot fight
-        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint32_t>,pairhash> shops;
+        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint16_t>,pairhash> botsFight;
+        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint16_t>,pairhash> botsFightTrigger;//trigger line in front of bot fight
+        std::unordered_map<std::pair<uint8_t,uint8_t>,std::vector<uint16_t>,pairhash> shops;
         std::unordered_set<std::pair<uint8_t,uint8_t>,pairhash> heal;
         const CatchChallenger::MonstersCollisionValue *monstersCollisionValue;
